# Code/Api/wwwroot/swaggerFastApi/v1/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "FastApi",
    "version": "v1"
  },
  "paths": {
    "/api/Execute/Process": {
      "post": {
        "tags": [
          "Execute"
        ],
        "operationId": "Process",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProcessRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProcessRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProcessRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiReturn"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiReturn"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiReturn"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiReturn": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProcessRequest": {
        "type": "object",
        "properties": {
          "processName": {
            "type": "string",
            "nullable": true
          },
          "parameters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/data"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "data": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "type": "string",
            "nullable": true
          },
          "info": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}